client: Move client logic to single module
